+++ client/examples/examples/bessel.hpp
@@ -13,9 +13,7 @@
 
 #include "autd3.hpp"
 
-using autd::NUM_TRANS_X;
-using autd::NUM_TRANS_Y;
-using autd::TRANS_SIZE_MM;
+using autd::NUM_TRANS_X, autd::NUM_TRANS_Y, autd::TRANS_SIZE_MM, autd::PI;
 
 inline void BesselTest(const autd::ControllerPtr& autd) {
   autd->SetSilentMode(true);
@@ -23,7 +21,7 @@ inline void BesselTest(const autd::ControllerPtr& autd) {
   const auto m = autd::modulation::SineModulation::Create(150);  // 150Hz AM
   autd->AppendModulationSync(m);
 
-  const auto center = autd::Vector3(TRANS_SIZE_MM * ((NUM_TRANS_X - 1) / 2.0), TRANS_SIZE_MM * ((NUM_TRANS_Y - 1) / 2.0), 0);
-  const auto g = autd::gain::BesselBeamGain::Create(center, autd::Vector3::UnitZ(), 13.0 / 180.0 * M_PI);
+  const auto center = autd::Vector3(TRANS_SIZE_MM * ((NUM_TRANS_X - 1) / 2.0f), TRANS_SIZE_MM * ((NUM_TRANS_Y - 1) / 2.0f), 0);
+  const auto g = autd::gain::BesselBeamGain::Create(center, autd::Vector3::UnitZ(), 13.0f / 180.0f * PI);
   autd->AppendGainSync(g);
 }