+++ client/examples/examples/stm.hpp
@@ -13,9 +13,7 @@
 
 #include "autd3.hpp"
 
-using autd::NUM_TRANS_X;
-using autd::NUM_TRANS_Y;
-using autd::TRANS_SIZE_MM;
+using autd::NUM_TRANS_X, autd::NUM_TRANS_Y, autd::TRANS_SIZE_MM, autd::PI;
 
 inline void STMTest(const autd::ControllerPtr& autd) {
   autd->SetSilentMode(true);
@@ -23,11 +21,11 @@ inline void STMTest(const autd::ControllerPtr& autd) {
   const auto m = autd::modulation::Modulation::Create(255);
   autd->AppendModulationSync(m);
 
-  const auto center = autd::Vector3(TRANS_SIZE_MM * ((NUM_TRANS_X - 1) / 2.0), TRANS_SIZE_MM * ((NUM_TRANS_Y - 1) / 2.0), 150);
-  const auto radius = 30.0;
+  const auto center = autd::Vector3(TRANS_SIZE_MM * ((NUM_TRANS_X - 1) / 2.0f), TRANS_SIZE_MM * ((NUM_TRANS_Y - 1) / 2.0f), 150.0f);
+  const auto radius = 30.0f;
   const auto point_num = 200;
   for (auto i = 0; i < point_num; i++) {
-    const auto theta = 2 * M_PI * i / point_num;
+    const auto theta = 2 * PI * i / point_num;
     const auto pos = center + radius * autd::Vector3(cos(theta), sin(theta), 0.0);
     const auto g = autd::gain::FocalPointGain::Create(pos);
     autd->AppendSTMGain(g);