+++ client/include/sequence.hpp
@@ -3,7 +3,7 @@
 // Created Date: 01/07/2020
 // Author: Shun Suzuki
 // -----
-// Last Modified: 26/12/2020
+// Last Modified: 27/12/2020
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2020 Hapis Lab. All rights reserved.
@@ -12,11 +12,19 @@
 #pragma once
 
 #include <vector>
+#include <memory>
 
-#include "core.hpp"
+#include "autd_types.hpp"
 #include "geometry.hpp"
 
 namespace autd {
+
+namespace sequence {
+class PointSequence;
+}
+
+using SequencePtr = std::shared_ptr<sequence::PointSequence>;
+
 namespace sequence {
 /**
  * @brief PointSequence provides a function to display the focus sequentially and periodically.
@@ -38,24 +46,6 @@ class PointSequence {
   /**
    * @brief Generate PointSequence with control points.
    */
-  static SequencePtr Create(const std::vector<utils::Vector3>& control_points) noexcept;
-  /**
-   * @brief Append control point
-   * @param[in] point control point
-   * @details The maximum number of control points is 2000.
-   */
-  void AppendPoint(const utils::Vector3& point);
-  /**
-   * @brief Append control points
-   * @param[in] points control points
-   * @details The maximum number of control points is 2000.
-   */
-  void AppendPoints(const std::vector<utils::Vector3>& points);
-
-#ifdef USE_EIGEN_AUTD
-  /**
-   * @brief Generate PointSequence with control points.
-   */
   static SequencePtr Create(const std::vector<Vector3>& control_points) noexcept;
   /**
    * @brief Append control point
@@ -69,7 +59,6 @@ class PointSequence {
    * @details The maximum number of control points is 2000.
    */
   void AppendPoints(const std::vector<Vector3>& points);
-#endif
 
   /**
    * @return std::vector<Vector3> Control points of the sequence
@@ -83,21 +72,21 @@ class PointSequence {
    * 1. The maximum number of control points is 40000.
    * 2. the sampling interval of Control Points is an integer multiple of 25us.
    * 3. (number of Control Points) x (sampling interval) must be less than or equal 1 second.
-   * @return double Actual frequency of sequence
+   * @return Float Actual frequency of sequence
    */
-  double SetFrequency(double freq);
+  Float SetFrequency(Float freq);
   /**
-   * @return double Frequency of sequence
+   * @return Float Frequency of sequence
    */
-  [[nodiscard]] double frequency() const;
+  [[nodiscard]] Float frequency() const;
   /**
    * The sampling period is limited to an integer multiple of 25us. Therefore, the sampling frequency is 40kHz/N.
-   * @return double Sampling frequency of sequence
+   * @return Float Sampling frequency of sequence
    */
-  [[nodiscard]] double sampling_frequency() const;
+  [[nodiscard]] Float sampling_frequency() const;
   /**
    * The sampling frequency division means the "(sampling period)/25us".
-   * @return double Sampling frequency division
+   * @return Float Sampling frequency division
    */
   [[nodiscard]] uint16_t sampling_frequency_division() const;
 
@@ -122,17 +111,7 @@ class CircumSeq : PointSequence {
    * @param[in] radius Radius of the circumference
    * @param[in] n Number of the control points
    */
-  static SequencePtr Create(const utils::Vector3& center, const utils::Vector3& normal, double radius, size_t n);
-#ifdef USE_EIGEN_AUTD
-  /**
-   * @brief Generate PointSequence with control points on a circumference.
-   * @param[in] center Center of the circumference
-   * @param[in] normal Normal vector of the circumference
-   * @param[in] radius Radius of the circumference
-   * @param[in] n Number of the control points
-   */
-  static SequencePtr Create(const Vector3& center, const Vector3& normal, double radius, size_t n);
-#endif
+  static SequencePtr Create(const Vector3& center, const Vector3& normal, Float radius, size_t n);
 };
 }  // namespace sequence
 }  // namespace autd