+++ client/include/twincat_link.hpp
@@ -3,7 +3,7 @@
 // Created Date: 01/06/2016
 // Author: Seki Inoue
 // -----
-// Last Modified: 26/12/2020
+// Last Modified: 27/12/2020
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2016-2020 Hapis Lab. All rights reserved.
@@ -14,7 +14,6 @@
 #include <memory>
 #include <string>
 
-#include "core.hpp"
 #include "link.hpp"
 
 namespace autd::link {