+++ client/lib/controller_impl.hpp
@@ -53,11 +53,11 @@ class AUTDController final : public Controller {
   void AppendModulationSync(ModulationPtr mod) override;
   void AppendSTMGain(GainPtr gain) override;
   void AppendSTMGain(const std::vector<GainPtr>& gain_list) override;
-  void StartSTModulation(double freq) override;
+  void StartSTModulation(Float freq) override;
   void StopSTModulation() override;
   void FinishSTModulation() override;
   void AppendSequence(SequencePtr seq) override;
-  FirmwareInfoList firmware_info_list() override;
+  std::vector<FirmwareInfo> firmware_info_list() override;
 
  private:
   std::unique_ptr<AUTDControllerSync> _sync_cnt;