+++ client/lib/ec_config.hpp
@@ -3,7 +3,7 @@
 // Created Date: 21/02/2020
 // Author: Shun Suzuki
 // -----
-// Last Modified: 25/12/2020
+// Last Modified: 27/12/2020
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2020 Hapis Lab. All rights reserved.
@@ -11,10 +11,9 @@
 
 #pragma once
 
-#include "consts.hpp"
-
+namespace autd {
 constexpr size_t HEADER_SIZE = 128;
-constexpr size_t EC_OUTPUT_FRAME_SIZE = autd::NUM_TRANS_IN_UNIT * 2 + HEADER_SIZE;
+constexpr size_t EC_OUTPUT_FRAME_SIZE = 626;
 constexpr size_t EC_INPUT_FRAME_SIZE = 2;
 
 constexpr uint32_t EC_SM3_CYCLE_TIME_MICRO_SEC = 1000;
@@ -30,3 +29,4 @@ constexpr size_t EC_DEVICE_PER_FRAME = 2;
 constexpr size_t EC_FRAME_LENGTH = 14 + 2 + (10 + EC_OUTPUT_FRAME_SIZE + EC_INPUT_FRAME_SIZE + 2) * EC_DEVICE_PER_FRAME + 10;
 constexpr double EC_SPEED_BPS = 100.0 * 1000 * 1000;
 constexpr double EC_TRAFFIC_DELAY = EC_FRAME_LENGTH * 8 / EC_SPEED_BPS;
+}  // namespace autd