+++ client/lib/emulator_link.cpp
@@ -3,7 +3,7 @@
 // Created Date: 29/04/2020
 // Author: Shun Suzuki
 // -----
-// Last Modified: 26/12/2020
+// Last Modified: 27/12/2020
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2020 Hapis Lab. All rights reserved.
@@ -19,13 +19,15 @@
 
 #include <algorithm>
 #include <bitset>
+#include <cstring>
 #include <iostream>
 #include <memory>
 #include <string>
 #include <utility>
 
+#include "autd_types.hpp"
 #include "consts.hpp"
-#include "geometry.hpp"
+#include "link.hpp"
 
 namespace autd::link {
 
@@ -82,7 +84,7 @@ bool EmulatorLink::is_open() { return _is_open; }
 
 void EmulatorLink::SetGeometry() {
   auto geometry = this->_geometry;
-  const auto vec_size = 3 * sizeof(Vector3) / sizeof(double) * sizeof(float);
+  const auto vec_size = 3 * sizeof(Vector3) / sizeof(Float) * sizeof(float);
   const auto size = geometry->num_devices() * vec_size + sizeof(float);
   auto buf = std::make_unique<uint8_t[]>(size);
   float header{};